Add some functions to compute areas and mean intensities for contours
Add functions for contour areas and mean intensities based on openCV
functions. The mean intensities computed suggest potential issues with
the thresholding process since the mean intensities range from both
above and below the kmeans threshold, which does not make a lot of sense
in terms of my understanding of the threshoding process. It is also
possible that the intensity calculations are flawed.