Add initial fourier filter
I've implemented the retreival algorithm, but I think it's fairly
sensitive to accurate physical parameters for the refractive index and
linear attenuation coeffcient. I can't even reproduce the projected
thickness for the latex spheres from the paper, which is troubling. It's
also possible, although probably easier to diagnose, that I've
implemented the filtering incorrectly.