Modify backgroundAnalysis
I'm not sure where the images in /test_images/ originated. The image
I've been working with, in my individual developer folder, is identical
to that found in the normalizedBackgroundRemoved folder in the drive.
But its shape and pixel values seem to disagree with those from the
/test_images/ folder, which is weird. I'd probably suggest replacing the
test_images objects with stuff taken from the drive folder. Also, I
looked at the old group's code, and it seems to do what it should. It
calculates a global max and min and scales everything according to
those values.