Add watershed methods to Shape class
Add methods to break droplets into smaller pieces using watershed method. Also, add a few lines
to the example code in iterative thresholding, and add an additional return value of the
original image. It might be better if the iterativeThresholding function
manipulated the copy of the original image rather than vice versa, but it's not a big deal for
me at this point. Finally, add a note that doesn't provide any information, and a copy of the
Shape class with all my print statements that I used while debugging the watershed methods to
my individual developer folder.