Add untested shape splitting criteria and splitter method
Add methods to shape class to remove small shapes that can be generated
by the watershed method. Recursively remove shapes and apply the
watershed method. The current method overwrites the original watershed,
although this data is recoverable using an overloaded pullSeeds and the
usual watershed and divideShapes. Although the methods are untested,
aside from checking to see that the childPad method returns a pad with
the correct shape, the new code isn't currently called anywhere, so I
don't think it'll hurt anything to commit them now. I'm really trying to
give a feel for how I plan to remove small shapes